feat(1378): add job tag
